+++ dbms/src/test/java/org/polypheny/db/cypher/Functions/TemporalFunTest.java
@@ -38,13 +38,7 @@ public class TemporalFunTest extends CypherTestTemplate {
 
     }
 
-    @Test
-    public void dateeFunTest() {
 
-        GraphResult res = execute( "RETURN DATE('2023-05-18')\n" );
-
-
-    }
